backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.6.0
django-extensions==4.1

